srcs/

+++ srcs/shared/do_pa.c
@@ -6,7 +6,7 @@
 /*   By: xmethula <xmethula@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/15 17:19:26 by xmethula          #+#    #+#             */
-/*   Updated: 2019/08/22 07:54:14 by xmethula         ###   ########.fr       */
+/*   Updated: 2019/08/28 07:47:30 by xmethula         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 void    do_pa(t_stack **stack_a, t_stack **stack_b, int flag)
 {
-    long int    tmp;
+    int     tmp;
 
     if ((*stack_b) != NULL)
     {
