+++ srcs/shared/do_rrb.c
@@ -6,7 +6,7 @@
 /*   By: xmethula <xmethula@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/22 08:19:35 by xmethula          #+#    #+#             */
-/*   Updated: 2019/08/22 08:20:52 by xmethula         ###   ########.fr       */
+/*   Updated: 2019/08/28 07:49:14 by xmethula         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 void    do_rrb(t_stack *stack, int flag)
 {
     t_stack     *p;
-    long int    tmp;
+    int         tmp;
 
     p = stack;
     if (stack->next != NULL)
