+++ srcs/shared/stack_add_top.c
@@ -6,14 +6,14 @@
 /*   By: xmethula <xmethula@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/15 17:01:07 by xmethula          #+#    #+#             */
-/*   Updated: 2019/08/19 08:49:36 by xmethula         ###   ########.fr       */
+/*   Updated: 2019/08/28 07:52:22 by xmethula         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../libft/libft.h"
 #include "../../includes/push_swap.h"
 
-void    stack_add_top(t_stack **stack, long int num)
+void    stack_add_top(t_stack **stack, int num)
 {
     t_stack     *new;
     t_stack     *tmp;
